Rip out 'long' and 'L'-suffixed integer literals. (Rough first cut.)

+++ Lib/base64.py
@@ -129,7 +129,7 @@ _b32alphabet = {
 _b32tab = _b32alphabet.items()
 _b32tab.sort()
 _b32tab = [v for k, v in _b32tab]
-_b32rev = dict([(v, long(k)) for k, v in _b32alphabet.items()])
+_b32rev = dict([(v, int(k)) for k, v in _b32alphabet.items()])
 
 
 def b32encode(s):
