+++ Lib/string.py
@@ -374,7 +374,7 @@ def rfind(s, *args):
 # for a bit of speed
 _float = float
 _int = int
-_long = long
+_long = int
 
 # Convert string to float
 def atof(s):
